:sparkles: Terminada la revisión del archivo y el enviode mensajes :sparkles:

+++ Prepa57/generarToken.py
@@ -4,7 +4,7 @@ import os
 
 # Archivos
 CREDENTIALS_PATH = r"c:\Users\Laura\Desktop\Data Scientist\Ejercicios Google Drive\secrets\credentials.json"   # tu archivo descargado de Google Cloud
-TOKEN_PATH = "token.json"               # aquí se guardará el token
+TOKEN_PATH = r"c:\Users\Laura\Desktop\Data Scientist\Ejercicios Google Drive\secrets\token.json"               # aquí se guardará el token
 SCOPES = [
     "https://www.googleapis.com/auth/spreadsheets.readonly",
     "https://www.googleapis.com/auth/drive.readonly"
@@ -14,19 +14,20 @@ def generar_token():
     if not os.path.exists(CREDENTIALS_PATH):
         raise FileNotFoundError(f"No se encontró {CREDENTIALS_PATH}")
     else:
-        return print("Ya existen credenciales validas")
-        
+        if not os.path.exists(TOKEN_PATH):
+            # Forzar login con refresh_token válido
+            flow = InstalledAppFlow.from_client_secrets_file(
+                CREDENTIALS_PATH, SCOPES
+            )
+            creds = flow.run_local_server(port=8080, access_type="offline", prompt="consent")
 
-    # Forzar login con refresh_token válido
-    flow = InstalledAppFlow.from_client_secrets_file(
-        CREDENTIALS_PATH, SCOPES
-    )
-    creds = flow.run_local_server(port=8080, access_type="offline", prompt="consent")
+            # Guardar token.json limpio
+            with open(TOKEN_PATH, "w") as token:
+                token.write(creds.to_json())
 
-    # Guardar token.json limpio
-    with open(TOKEN_PATH, "w") as token:
-        token.write(creds.to_json())
+            print(f"✅ Token generado correctamente en {TOKEN_PATH}")
+        else:
+            print(" 🎉 Ya existe token")
 
-    print(f"✅ Token generado correctamente en {TOKEN_PATH}")
-
-generar_token()
+if __name__ == "__main__":
+    generar_token()
